python_casssandra_redis_kafka/app/requirements.txt
redis
cassandra-driver
kafka-python
